requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
tensorflow>=2.16.0
keras>=2.16.0
numpy>=1.24.0
ocr-tamil==0.4.1
aksharamukha==2.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
openai==1.3.0
pypdf2==3.0.1
scikit-learn==1.3.2
